created add-to-project controller and popup
